requirements.txt
tensorflow==2.13.0
google-cloud-aiplatform==1.43.0
